Finished the form.
I stretched the width to 100% for that the input field stetches the length of the form. I made the form input display:block. I gave padding to the form input so the white space would be bigger. I added a 1px border to the input field to add emphasis to sign up. I increased the font size to 16 px on the form input. I colored the form input text to #024e76, I aslo gave the form input of 15px margin bottom so they can have space from each other.